deps: use custom build of aria-query@5.0.0
Version 5.0.0 of the aria-query package currently includes files with
spaces in their names [1], but those files aren't used at build- or
run-time. Since this package is always present in a hydrated
cockroach.git clone, a custom version of aria-query [2] can be used with a
yarn "resolutions" entry, i.e.:

    {
      "name": "example-package",
      "dependencies": {
        "...": "..."
      },
      "resolutions": {
        "aria-query": "file:../rel/path/to/yarn-vendor/aria-query-5.0.0-no-spaced-files.tgz"
      }
    }

Replace the corrupt aria-query build with a custom build that doesn't
include extra files.

[1] https://github.com/A11yance/aria-query/issues/367
[2] https://github.com/cockroachdb/aria-query/releases/tag/v5.0.0-no-extra-files