deps: add aria-query@5.0.1
aria-query@5.0.0 included extra files with spaces in their filename,
which caused errors when Bazel interacted with them. To work around
that, we produced a copy of 5.0.0 that removed those files [1]. The
upstream maintainers published version 5.0.1 which includes no other
changes (but no spurious files) [2], so add that to the set of available
dependencies.

[1] 1dea01c (deps: use custom build of aria-query@5.0.0 (#95), 2022-06-15)
[2] https://github.com/A11yance/aria-query/issues/367